\name{EMandMMPackage-package}
\alias{EMandMMPackage-package}
\alias{EMandMMPackage}
\docType{package}
\title{
\packageTitle{EMandMMPackage}
}
\description{
\packageDescription{EMandMMPackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EMandMMPackage}
\packageIndices{EMandMMPackage}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{EMandMMPackage}

Maintainer: \packageMaintainer{EMandMMPackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
